Implement Support for P2E Ability Boosts
Boosts add +2 to abilities < 18, or +1 to abilities >= 18. They can also
be stacked.

Content/OpenPathfinder2e/Core/GameEffect/GE_BoostAbilities right now is
a demo of how to apply them to a character. Ideally we should have a
more flexible Blueprint or something we can supply parameters/values to
that can handle this instead.