Denote How We are Complying with OGL Section 8
According to the (ancient) WotC Software FAQ (v1.0 -
http://www.wizards.com/default.asp?x=d20%2Foglfaq%2F20040123i):

--- snip ---
Q: How can the OGL be used with software?

A: Just like with other material, the OGL allows you to use any Open
Content, provided you follow the terms of the OGL. Follow the
requirements of the License, include the text of the license and the
appropriate copyright information, and clearly identify Open Content.

NOTE: The biggest problem we've found with software and the OGL is that
programmers aren't paying attention to Section 8 of the OGL. Section 8
states: "If you distribute Open Game Content You must clearly indicate
which portions of the work that you are distributing are Open Content."
This doesn't mean you can say "all rules in my program are Open", the
users need to be able to see all that Open Content. You can do this by
putting Open Content in a format that is easy to understand. Popular
solutions have been ... using Java script on a webpage (viewing the
source of the webpage will display the code and Java script is relatively
easy for a user to interpret). The key is that the user has to see
everything that is Open Content that the program uses and be able to
understand it without too much effort. The whole point of the OGL is
that once information is declared Open everyone has free access to it
under the OGL. Compiling that information into a program denies the user
that access and violates the spirit of the Open Gaming License.
--- snip ---

To satisfy this requirement, we're using inline citations in comments
and keeping the relevant code short and straightforward. If that is not
sufficient, too bad -- the license literally says NOTHING WHATSOEVER
about source code forms and understandability, only CLARITY of
INDICATION.