Abandon Execution-based Approach for Ability Boosts; Use GEs Instead
This reverts to using MMCs for calculating each boost. To accomplish
what we were doing with the ExecCalcs, we can instead chain GEs that
increment both the ability and the counter. That is a much more robust
approach than ExecCalcs since we can guard boosts by required tags (as
in the case of the key ability boost GE) and the MMCs support
prediction.