Make PF2 Characters Auto-load Core GEs; Control Order by Weight
This makes it significantly easier to setup a PF2 character by ensuring
that all the GEs that calculate core stats are automatically loaded in
the ideal order, while still allowing custom GEs to be loaded in the
middle of the list.

We may want to expose the weight options to the blueprint editor, but
for now all "additional" GEs are assigned the same order and evaluated
in the order that the GEs were specified by the game developer.