Make it Possible to Track Known Lore Names
This isn't all that useful for a video game as it tends to be used more
by DMs for tabletop RPGs. We also have a limitation with skills since we
only support up to 2 lore skills (just like on official character
sheets) but apparently characters can learn any number of lore skills.