Code Clean-up -- Put Properties of PF2CharacterAbilityBoostCount First
Trying to make the organization of this struct consistent with the way
we do things in classes. We are using getters and mutators here instead
of making properties public because we want to apply business logic to
the decrement of counts.

We are using a struct since it's lighter-weight and Blueprint works
better with structs when exposing them as properties of the referencing
PF2CharacterBase.