Make Boost GAs Functional + Replace Manual Ab. Boost Counts on Charac.
Now, ability boosts are granted/given to characters just like regular
GAs, but we surgically apply them by matching-up Ability Boost
Selections.

To streamline and unify the way that ability boosts are applied to
characters by both game designers and the player, we no longer have
a property for setting boosts by count. Instead, we expose a property on
the character for specifying what boosts have been selected for a
particular type of boost GA. If the character has that boost GA as one
of their granted abilities, then the boost selection gets "consumed" to
activate the corresponding GA.

This eliminates the need for a callback into the character base class to
update boost counts, so the IPF2CharacterAbilityEventsInterface is no
longer needed.