Implement First Working Ability Boost Modal Prompt
It doesn't yet get dismissed when the user has completed their choices,
and it doesn't allow the user to "undo" the choices they've made, but
by golly if it ain't pretty to see an ability boost get rendered on
screen and to have the various buttons react to what choices are valid.

Each row of the dialog is dynamic, driven off of the possible ability
score types.