Major Re-factor -- Track GE Weights as FNames + Tags; Re-apply Later GEs
Now, when a passive GE is applied to an ASC that already has passive GEs
activated, we only refresh GEs in "weight groups" that appear after the
weight group to which the passive GE is being added.

In addition, the weight of a GE can now be controlled by a tag on the GE
so that a game designer can set it instead of it having to be
hard-coded.