Ensure that First GE in a Weight Group Gets Activated
Without this, the first ability boost added to an ASC while passive GEs
are already applied would end up not activated. Consequently,
re-applying subsequent GEs would have no effect.

Instead, we assume that if passive GEs are active and this is the first
passive GE in a particular weight group then we should activate the new
GE. If the caller doesn't want this behavior they should not be adding
the GE while passive GEs are active on the ASC (they can simply toggle
them all off first).