Fix/Re-factor Passive GE Group Activation and Deactivation; Add Caching
Previously, multiple places in the code were manipulating the list of
activated effects and doing so inconsistently. Now we do this through
dedicated methods.

In addition, the list of passive GEs to apply is now cached so that we
avoid having to recompute the values every time they are needed.