Expose (Activate/Deactivate)PassiveGameplayEffects Methods to Blueprint
We want game designers to rely on these methods for toggling passive GEs
on a character, so let's make them public to encourage their use.

Generally, though, passive GEs are automatically activated when the
character is first possessed by a player controller and they are usually
re-activated when the character levels up. So, game designers may not
actually need to deactivate passive GEs manually in Blueprint that
often unless the game has some very specific requirements outside the
norm here.