Rename `UPF2GameplayAbility_BoostAbilityBase` to `UPF2AbilityBoostBase`
This was the only situation where we were prefixing a base class type
with the name of its parent class, leading to a very verbose, very long
name. Now, the name is shorter and easier to grok.

For now, the corresponding target data class for this object will remain
as `FPF2GameplayAbilityTargetData_BoostAbility` until we either come up
with a better class naming convention or start using namespaces in a way
that makes that name unnecessarilty long.