Add PF2 Version of "Play Montage and Wait for Event" Ability Task
This makes it much easier for us to play a montage that may return any
one of several results (in the form of event tags).

This is a cleaned-up version of an ability task with the same name from
the Action RPG sample project. We should be safe to pull this into the
project (from a licensing perspective), since the content carries the
following note:

> UE-Only Content - Licensed for Use Only with Unreal Engine-based
> Products

From:
https://www.unrealengine.com/marketplace/en-US/product/action-rpg