Move TEML Calculation into `Calculations/` Folder
Since this calculates a value, it makes sense to put it with the other
calculations. But, most of what is in the `Calculations/` folder are
a mix of MMCs and executions; so -- at some point -- this is likely
going to get re-organized and refined even further to add more rhyme and
reason to what goes where.