[#12] Move EPF2ModeOfPlay into Separate File; Move Game Mode Interfaces
The EPF2ModeOfPlay enum gets used a lot now, and it's confusing that
it's part of the same include file as the game state. So, let's not
continue that confusion.

Separately, future work will add additional game mode rule set classes
so it makes sense to break out the game mode interfaces and related code
into their own folder.