[#12] Implement "Mode of Play Rule Sets" (MoPRS)
MoPRS not only provide logic that control how different gameplay events
are handled, but also can act as an extension to game state by storing
and maintaining variables that are relevant for the current mode of
play. For example, encounter modes maintain initiative order, number of
enemies left standing, etc. which are not relevant in other game modes
like exploration mode. A new MoPRS instance is created each time that
the mode of play changes, so this state is only maintained while it is
relevant.