[#12] Rename `StartXMode` to `RequestXMode` and Make Public
These methods are subject to the rules of the game mode, so we don't
need to worry about them being public. At the same time, calling them
`StartXMode` could lead a game designer to think that their effect is
immediate, as that name implies that they are directing a mode switch
rather than requesting one if allowed by the rules.