[#7] Implements a Utility Method for TScriptInterface Conversions
This makes it a little easier to deal with parts of the API that expose
an interface object to Blueprint. This should be used sparingly, since
I expect this carries a performance penalty and it could lead to places
where an object that implements the interface without being a UObject
could accidentally leak out to Blueprint, resulting in a crash or
nullptr.