[#7] Implement Base Class for "Queued" MoPRS Logic
Blueprint has no concept of a MultiMap, so it would be tedious for an
MoPRS blueprint to maintain all of the different action queues for each
of the characters in an encounter. To make things easier, this new base
class provides convenient, Blueprint-callable methods for manipulating
action queues for each character.