[#7] Fix Client-side Prediction of WaitForInitiativeTurn Ability Task
Previously, an ability that was blocked waiting for initiative on the
server only executed on the server when its turn was up. This change
seems to fix that by signaling the client to simulate a callback to
finish off the ability task once the callback on the the server's copy
of the ability task comes through.

This approach was borrowed from UAbilityTask_NetworkSyncPoint.