[#7] Allow PF2 Game Mode to "Tick"
This is needed so we can pass ticks on to MoPRS without making each an
actor. There may be a performance penalty here but for now let's run
with it. Later there may be a way to eliminate ticking from MoPRS all
together, if we can make sure they have the opportunity to react to
time-based events on their own.