[#7] WIP Encounter Mode MoPRS PoC
This is just a proof of concept that allows the actions of the first
player in a networked match to have their actions queued up and then
executed when that player's turn comes up. Right now, initiative rolling
logic isn't part of the MoPRS -- we're literally just using ticks to
execute any queued actions every 6 seconds. This will need quite a bit
more work to get this to match PF2 rules.

Ideally, the "cost" for each action would come from an action points
stat on the character that gets reset every time they start a turn. When
played via tabletop -- or if we were forcing OpenPF2 to be used only
with turn-based games -- then this would be straightforward. But, since
this is intended for use with Action RPGs, we need to consider that a
player's turn ends either when they're out of action points OR after six
seconds, to keep the combat loop exciting and fluid. So that means there
are two events that can trigger a turn change -- points and time.