[#7] [#12] Implement Initiative Rolling and Interaction with Encounter MoPRS
This implements a default initiative-rolling GA that rolls against the
character's perception. This relies on the Encounter MoPRS adding the
desired initiative-rolling GA to the character and then listening for a
callback in the form of a gameplay event tag. That allows the MoPRS to
control how initiative is rolled simply by picking which roll GA is
used.