[#7] [#12] Abandon GA-based Initiative; Move Logic into Encounter MoPRS
Since the MoPRS determined what GA would be added to the character to
determine their initiative and the MoPRS is ultimately what needed the
initiative value, it seemed silly to even use a GA for this. Really, all
the control is in the MoPRS so it makes way more sense to move the logic
there.

That also eliminates any delay in obtaining the initiative in the
MoPRS and makes it easier for game designers to setup encounters that
use different skills for initiative rolls -- all they need to do is
extend the base MoPRS blueprint with a different "base skill" value
supplied in the class defaults settings.