[#7] [#12] WIP First Working Initiative and Turn Proof of Concept
This correctly cycles between characters every six seconds, executing
any actions that the character has queued up. The encounter MoPRS
handles setting up each of the characters with initiative and allows
each of the characters to take turns according to their initiative.

Only major issue with this approach is that all of the queued-up
abilities execute at once rather than in sequence. We'll need to find
some way to block execution of each ability until the previous one has
finished.