[#7] [#12] Move Blueprint Implementable Events into Base Classes
`BlueprintImplementable` events are really more of an implementation
detail rather than something callers care about. Additionally, if we
specify them in an interface, then UE4 requires the interface to be
Blueprint-implementable, but then we can't have any `BlueprintCallable`
functions specified in the interface.

To address these issues, the interfaces that trigger BP-implementable
events now specify `HandleX` methods that are implemented by the base
classes to dispatch the `BlueprintImplementable` events. BPs that want
to be able to support those events must extend from the base classes,
but the interfaces themselves are no longer BP-implementable.