[#7] Fix Propagation of Action Queued/Dequeued Events from Interfaces
Looks like these have to be declared multicast so that they get
triggered on the server but execute on the client. Separately, though
some resources online seemed to suggest that we should not repeat the
UFUNCTION macros in sub-classes, it seems that this may be incorrect or
at least if it is not done for RPCs, they are not properly setup in
sub-classes.