[#7] [#12] Fix Issue with Incomplete `AGameStateBase` Type
For some reason, recent changes started to cause builds to fail,
complaining that `AGameStateBase` was incomplete, citing the
`GetGameState` call in `GetModeOfPlayRuleSet()`. Although possibly an
issue with header includes order or some other dependency issue, the
most straightforward fix was to move this previously-inlined method
into being a regular method that is no longer declared in the header
file.