[#7] [#12] Introduce `PF2QueuedActionHandle` and Replicate That Instead
The previous approach of passing around a queued action into Character
and PlayerState multicast callbacks wasn't smart; the action was being
referenced by a pointer that's only going to exist on the server!

This changes the approach to using a handle struct that provides the
name, image, and a unique handle ID of the real action so that clients
can still reference the action without having to replicate it.