[#7] Introduce `GetCharacterIdForLogs()` on PF2 Character Interface
Trying to debug multi-client, multi-thread issues is a pain in the ass
without some way of distinguishing between multiple characters that
might have the same name as set by the game designer. This new method
makes it easy for logger calls to get a unique ID for each character,
beyond just the character's name.