[#7] Standardize Exposing Log IDs via PF2LogIdentifiable Interface
This makes it easier to log the IDs of objects as more and more PF2
objects add support for this interface.

Sadly, because of an obscure bug in the way interfaces work in UE, it's
not possible for IPF2CharacterInterface to extend from the new interface
because it already extends from a different interface and interfaces do
not properly handle multiple inheritance. But... the actual
implementation class can inherit from both! So... we just duplicate the
methods in both interfaces.