[#12] WIP: Fix Transitions Out of Encounter Mode
Previously, the logic was flawed and only allowed transitions from one
encounter to another even though that's not what we want (we want the
opposite). Now, we block encounter-to-encounter transitions and allow
other transitions only when all foes are vanquished.

Still need to implement/fix logic to handle clearing action queues on
the way out of the encounter.