[#21] Track which PF2 Character(s) are Controlled, Separately from Pawn
During encounters, the PC will be "possessing" a camera pawn rather than
any of the party members, so we cannot rely on the Pawn to be the source
of record for what characters a particular PC can actually control. In
addition, for #22 we'll need to implement party management which will
presumably allow the same PC to control multiple characters and all will
need to participate in the encounter even though only one can be
controlled at a time.