[#23] Implement Interfaces for Commands, Command Queues, and Queue Comp.
This will replace what we're currently doing in the encounter MoPRS base
class. We are going to be moving the action queueing for each character
out of the MoPRS base and into a component within each character, giving
us more flexibility for each character to manage their own queue of
commands/actions. In addition, each ability will be wrapped in a
Character Command, allowing us more flexibility in when the ability is
actually getting executed.