Relax ASC Type Requirements for Characters
We have a PF2-specific ASC interface, so we should not require a
specific base class for character ASCs. At the same time, we can't use
the interface as the type of this field because the interface doesn't
include all the base ASC object methods (ASCs don't have a base
interface). So, we have to treat this like a generic ASC and case it to
our PF2 interface, but otherwise this leaves the door open for
implementors to use their own custom character ASCs without having to
extend from ours.