[#23] Implement Multicast Delegate Callbacks on Command Queue Components
This should allow the UI to be notified when commands are queued or not
for the active character without our having to dispatch callbacks
manually to the player and player controller as we had previously done.

It's a shame that this can't be done on a UINTERFACE, but it appears
that this is a limitation of interfaces as delegates are an
implementation detail. Technically, I have read that it's possible to
define methods in the interface that manually bind the delegates behind
the scenes but that feels non-obvious enough as to confuse most game
devs. Typically, if something needs to bind an event to a component, it
can be fairly confident in exactly what type of component it is.