[#23] Move PC Convenience Methods Out of MoPRS Base and into New Library
I don't really like having to move this into a utility library because
it feels like something that belongs in a base class or some other
object, but there are several places in the code that need to access
this logic and copying-and-pasting this behavior feels like an odd fit
here. In addition, we're likely to want to invoke this from Blueprint,
so it makes sense as a library function.