[#23] Move Character Initiative Queue Behaviors into Dedicated Component
This allows the logic for maintaining queues of characters by initiative
to be re-used, and reduces the need for encounter MoPRS implementations
to have to extend from `UPF2EncounterModeOfPlayRuleSetBase`. At the same
time, `UPF2EncounterModeOfPlayRuleSetBase` is now a facade, providing
several convenient, BP-callable methods that call into the appropriate
components so that implementations do not need to be coupled to specific
components directly in each Blueprint.