[#23] WIP Model Encounter MoPRS Business Logic Using Behavior Trees
This creates a new behavior tree with suporting services and decorators
for executing the business logic of the encounter Mode of Play Rule Set
(MoPRS). This vastly simplifies what was previously a very confusing and
error-and-race-condition-prone, time-driven blueprint.

This isn't yet managing characters in the encounter correctly, but it's
a good start.