[#8] [#23] WIP Dynamic Bindings from Input -> Commands
This implements a new actor component, `UPF2CommandBindingsComponent`,
which can be added to playable PF2 characters to allow them to track
which abilities are bound to specific controls. The component should
game designers to bind to abilities both by GA class (for custom
abilities not granted by ancestry and class GAs) as well as to bind by
GA handle (for abilities granted by ancestry and class).

Right now, the former use case is not yet working properly because of
differences between client and server authority (e.g., the client needs
the handle of the ability that can only be granted on the server).