[#22] [#23] WIP Party Management (Borrowing from UE4 RTS)
This introduces a new `PF2OwnerComponent` that can be added to character
actors to track the party and player that they belong to. This also adds
tracking of party affiliations for players as well.

This approach was borrowed from the UE4 RTS project, but modified to
work with parties instead of teams.

This code is NOT YET functional.