[#23] Fix-up `GA_StandardMeleeWeaponBase` Blueprint
Not sure when this broke, but the reference to "Play Montage and Wait
for Event" was totally broken. I am also not sure if we want to wait for
an event here, so for now I'm just making this a standard "Play Montage
and Wait" task call, with the montage being instance-editable. It does
not appear that this is used yet, so we can make more changes to this
later as needed.