[#23] WIP Assign PCs a Player Index + Use it to Auto-Claim Characters
This is borrowing a bit more from UE RTS to allow us to claim the
characters that make up a particular player's party when they join the
game. Not sure I like this pattern just yet b/c it's better suited for
an RTS, whereas in an RPG the party tends to be dictated by story
outside the map. For now, though, I'm rolling with it.