[#22] [#23] Make Owner Components Aware of Party/Remove "Claiming" by ID
This moves inter-player, inter-character logic that was previously in
the player state and ownership component code into the game mode. Now,
only the game mode has the responsibility of moving a player between
parties and of transferring ownership of a character from one player to
another.

Parties are now full-fledged, blueprintable objects that can have events
to respond to when a player gets added to the party. As with most things
in PF2, parties are backed by an interface.

The `PF2OwnerTrackingComponent` now no longer maintains the ID of the
initial player owner, as this was suitable for an RTS but isn't well
suited to an RPG use case (it's tricky, fragile, and error-prone to keep
track of which player ID should be assigned which characters). Instead
of placing characters in a map and hoping that they get assigned to the
proper players, it's better for us to specifically grab the desired
character in the game mode or somewhere else on the server and assign
ownership of that character explicitly. This use case is best handled
via calls into the game mode.