[#22] Move Controlled Character List from Plyr. State to Plyr. Control.
It feels more natural for the player controller (PC) to own the list of
characters that the player can take control of, rather than it being a
responsibility of the player state (PS). For one thing, the PC has to
react to receiving control of or releasing control of a character, which
it can do easily if it owns replication of the list; otherwise, we have
to notify the PC from the PS or use multicast event notifications, which
is less elegant. Secondly, I can't think of any reason why we need to
notify _all_ players when a specific character has been given over to
a particular player -- we only need to notify the affected PCs. So, the
PC is a more efficient choice.