[#8] [#21] [#22] [#23] Move Command Execution RPC Back into PCs
The player controller has to handle execution of player-issued commands
because it's the only reliable net authority that we can route commands
through on the client side. The command bindings component of a pawn
that is unpossessed can't send RPCs to the server on the client's
behalf without going through the PC for the player.

In addition, this change allows us to move responsibility for
constructing commands from the PC and AI controllers into the PC Server
RPC, simplifying the execution logic. The PC and AI controllers are
still asked to actually invoke the commands, to give them control over
the process for characters that they are possessing.

A command issued to an AI-possessed pawn that is controllable by the
local player would still get routed through the local player's PC to
the server where the command then gets dispatched to the possessing AI
controller. It's a bit roundabout, but I don't see another reliable way
to navigate net authority rules.