[#8] [#21] [#22] [#23] Move Character Command Creation to Server
Character Commands have to be created on the server side or they won't
be available for the server to act on them; actors spawned by a client
don't exist on the server. So, in hindsight, commands might not even
have to be actors at all, but I am leaving them as such for now to keep
options open for game developers in case they have a use-case for a
server-spawned command to be replicated to a client.

Since commands have to be spawned server-side, the RPC has to take in
the details of the command -- the handle and the target actor -- as
parameters and then create the command from those values. This is still
cleaner than the prior implementation where the command was created by
the PC or AI controller, so I'm all right with this.