[#22] [#23] Eliminate `CharacterControllerInterface` + Simplify Commands
Doing a 180 on the prior approach -- we don't actually want to route
character commands through the possessing controller at the time that
the command is being invoked or queued. Rather, AI controllers should
only be notified of a command when it is ready to be *executed* (either
immediately or after being dequeued).