[#23] WIP Move Active Encounter Character out of Blackboard into MoPRS
The active character should be a function of the Encounter MoPRS rather
than the Encounter AI, since even a non-AI Encounter MoPRS approach
would need to track which character should act.

This new approach doesn't quite work yet; it crashes out interacting
with the character command queue. More on that soon.

Also, the AI decorators and tasks should not be casting to the encounter
MoPRS base class -- we should move these function definitions to an
interface so that the Encounter MoPRS has more implementation
flexibility.