[#23] Rework Command Cancellation to Work on Clients with Replication
This ensures that clients can push requests up through their Player
Controller to de-queue a command for a character on the server side.
There may be a more elavorate way to do this so that commands can be
prediected client-side more readily, but for now this appears to work.

There is still a subtle bug that causes commands not to be cancelled if
the player attempts to cancel them too quickly, but I haven't yet been
able to nail that down.