[#23] Extract Encounter-specific MoPRS BlueprintCallables into Interface
A key design principle in OpenPF2 is that everything should use
interface methods to communicate, rather than assuming that any other
object extends from a specific base class. Here, the Encounter MoPRS
base class was providing a lot of functionality that gets leveraged by
the encounter BP and AI, but it was casting the MoPRS to a specific base
class, limiting the reusability of this logic for other projects.

Now, the Encounter MoPRS BP and the AI that runs encounters is using
interface methods for this, allowing them to be used with any encounter
MoPRS that implements the interface instead of one that must extend our
default base class.