[#23] Cite Source for Six Second Limit on Character Turns
I am happy that I went through the exercise to cite my source here,
because it seems that the entire ROUND is supposed to be limited to six
seconds, NOT each character's turn WITHIN the round. This will likely
require some thought and then later some additional logic in the
encounter AI and supporting tasks to pull off.

The main question I have is, if the entire round is supposed to last six
seconds, what happens if a character takes too long? How do we know if
the character is taking too long? Should we be dividing six seconds
across the number of characters in the encounter. If there are 100
characters in the encounter, then each character would have only 60 ms
to act, which is way too short. This remains to be decided.