[#22] Define Event Callbacks for Changes to Input Binding Connections
This makes it possible for the UI to indicate when a particular character has
been wired up to input. Unfortunately, this alone isn't enough for us to
indicate what character the player has currently selected since the in-game UI
UMG widget ALSO needs to monitor changes in selected character so that it can
adjust which command queue to display.