[#22] Implement `UPF2ActorLibrary::GetComponentByInterface()`
Surprisingly, there is no built-in already defined for getting only a
single component by its interface. This addresses that shortcoming to
make it possible for game designers to get components by interface in
much fewer steps. Without this, they have to use
`GetComponentsByInterface`, do a length check, grab the first element,
and then cast it.