[#22] Cache Active Char. in Queue Comp. and Expose Old Char to Listeners
Listeners to the `OnActiveCharacterChanged` events (Native or Blueprint)
almost always need to know who the previously active character was so
that input can be properly unbound from that character. This was not
possible in the earlier version of this component because we could only
look-up characters by index in the queue of characters, and the
character would be missing from the queue on "Remove" or "Clear"
actions. But... if we don't look up the active character by index each
time it is requested and instead cache the active character in a member
field, we're able to do proper character delta handling while still only
replicating the character index.