[#22] Remove Character Pointers from `PF2CommandInputBinding`
Each individual binding shouldn't have to care what character it's bound
to, and the bindings component should be smart enough to recognize what
character it's owned by. The previous interface was awkward because it
was conceivable that a different character than the owning character
could be passed-in to the `LoadAbilitiesFromCharacter()` method, which
could lead to some very strange side effects. In addition, by having
the binding have to keep track of the character, replication would be
made more awkward because we'd have to convert the reference into an
Actor.