[#22] Implement Support for Capturing Deltas with Arrays of Structs
Structs can't be directly compared; rather, we have to use some unique
key field within a struct to identify when another struct is equivalent
to it. This new method, `CaptureStructDeltas()`, takes in an equality
function that supplies this equality logic on a per element basis. The
older `CaptureDeltas()` method has been renamed `CapturePtrDeltas()` to
avoid confusion. Technically, we could just use the same name and make
them overloads since the argument lists are different, but conceptually
the two methods work very differently.