[#22] Add Optional Support for Replication of Command Bindings
This may be necessary for saved games to include binding mappings, but
at the moment this is just an optional feature. This was originally
added because it seemed like we needed to load character abilities on
the server side in order to populate the bindings array, but it turned
out that this was a red herring -- the character's ASC was just not
being initialized when the client was swapping to a character that had
not yet been active. So, for now, we don't need to replicate the
bindings from the server and can just rely on the client's copy of the
ASC to drive the default bindings.