[#22] Move Ability Management to Server + Replicate Bindings from There
It doesn't seem like there's a way to get bindings to work properly
unless they're replicated from the server. The problem is that at the
time that abilities are being loaded from the character, the local ASC
might not have replicated its `ActivatableAbilities` and so the client's
view of the abilities will be inaccurate. This wasn't a problem for the
initial character that gets possessed, but characters that the player
possesses later were ending up with a blank ASC to start.