[#22] Stop Replicating Command Bindings from the Server; Use Delegates
Replicating command bindings from the server copy of an OpenPF2 Command
Bindings component ensured that we could always load abilities from the
server copy of each character's ASC, but created the problem that any
time the server changed the bindings, we'd lose the local binding
context. In other words, if we bound an "Attack" input to a particular
character ability and saved the handle on the client copy of the
binding, but the character then gained a new ability later, we'd lose
track of the handle for the binding after the server replicated its new
copy of the character's ability bindings. Clearly, replication was not
the answer.

Instead, we now fire a dynamic multicast delegate any time that the
client copy of a character's abilities have changed on the ASC. The PC
can link up this delegate to call into the command bindings component
to reload bindings whenever the abilities in the character's ASC have
changed, allowing the client side to be fully in control of how the
existing bindings are unbound/re-bound.