[#21] Make it Possible to Find Granted Abilities and Handles by Tags
This allows the handle for a granted ability to be located by its tags,
such as when we need to bind an ability to a command that will be
executed in the future. Without this, we'd need to make the association
between a command and the ability its invoking based on gameplay event
tags, which would work but is a looser association that opens us up to
invoking more than one ability when we intend to only invoke one.

For now, stricter is better because there's no ambiguity about what
ability a command will invoke. But, later we may find commands need to
be more flexible by allowing them to trigger abilities by tag instead.
We shall see.