[#21] Clean-up Camera Pawn + Make Tilt Zoom Affect Client Transitions
There's more work to be done here because it seems very strange that the
camera transitions are server-driven rather than client-driven. But, at
least with these changes, a client can apply a tilt-zoom during a
transition and the transition will gracefully adapt rather than ignoring
the tilt-zoom until it "snaps" into the desired location and rotation at
the end of the transition.