[#21] Move All Camera Transitions to Client Side
By replicating the strategic camera pawn in the player controller, we no
longer need to control the transition camera or real camera from the
server side. This simplifies the logic for the transition camera since
we no longer need to worry that transitions are happening server side
while the player is adjusting tilt zoom client side.

Separately, this should reduce overhead on the server during camera
transitions into and out of encounter mode because the server is no
longer animating the cameras for multiple clients -- the clients do it
on their own.