[#21] Enable Bird's Eye Camera Pawns to Supply their Camera Components
Rather than assuming that the components have only one camera (or that
the camera component is a `UCameraComponent` at all), the pawn can now
provide the exact component it wants to designate as the camera for
tilt zoom and zooming capabilities. For example, this would allow the
pawn to return a spring arm component in place of the actual camera.