[#21] Add Automatic Ground Height Detection + Transition w/. Easing
Now, the bird's eye camera pawn will detect the height at which it
enters the scene and automatically adjust its height to maintain the
same relative distance from the ground even if the ground slopes up or
down (+/- 5 meters).

There may be more fine-tuning we can do here later, but with the
exception of invisible collision primitives that result in an unexpected
ground height result, this is looking pretty good in the OpenPF2
Playground sample map.